fix: content script fails to load
