+++ src/index.tsx
@@ -1,3 +1,4 @@
+import './provider.worker'
 import React from 'react'
 import { HashRouter as Router, Route, Link } from 'react-router-dom'
 import ReactDOM from 'react-dom'
@@ -29,9 +30,8 @@ import NearMe from '@material-ui/icons/NearMe'
 import Assignment from '@material-ui/icons/Assignment'
 import Phonelink from '@material-ui/icons/Phonelink'
 import { ExportData } from './components/MobileImportExport/Export'
-import { uiSetup } from './setup'
+import './setup.ui'
 
-uiSetup()
 const drawerWidth = 240
 const empty = (
     <div
