+++ src/provider.ui.ts
@@ -1,0 +1,1 @@
+import './social-network-provider/facebook.com/ui-provider'
