+++ src/provider.worker.ts
@@ -1,0 +1,2 @@
+import './social-network-provider/facebook.com/worker-provider'
+import './social-network-provider/options-page/index'
